Extract theme manager staging boundary
