+++ scripts/test-theme-manager.js
@@ -52,6 +52,7 @@ const {
 
 const themeManagerSource = readFileSync(new URL('../assets/js/theme-manager.js', import.meta.url), 'utf8');
 const themeManagerDataSource = readFileSync(new URL('../assets/js/theme-manager-data.js', import.meta.url), 'utf8');
+const themeManagerStagingSource = readFileSync(new URL('../assets/js/theme-manager-staging.js', import.meta.url), 'utf8');
 const themeManagerViewSource = readFileSync(new URL('../assets/js/theme-manager-view.js', import.meta.url), 'utf8');
 const themePackageCoreSource = readFileSync(new URL('../assets/js/theme-package-core.js', import.meta.url), 'utf8');
 const themeInstallServiceSource = readFileSync(new URL('../assets/js/theme-install-service.js', import.meta.url), 'utf8');
@@ -200,15 +201,22 @@ await run('keeps theme package and install responsibilities outside the UI contr
   assert.match(themeManagerSource, /from '\.\/theme-package-core\.js';/, 'theme manager should consume package rules from the core module');
   assert.match(themeManagerSource, /from '\.\/theme-install-service\.js';/, 'theme manager should consume install staging through the service module');
   assert.match(themeManagerSource, /from '\.\/theme-manager-data\.js';/, 'theme manager should consume registry, catalog, and product-state loading through the data module');
+  assert.match(themeManagerSource, /from '\.\/theme-manager-staging\.js';/, 'theme manager should consume theme staging through the staging module');
   assert.match(themePackageCoreSource, /export function collectThemeArchiveEntries/, 'theme package core should own ZIP analysis');
   assert.match(themePackageCoreSource, /export function normalizeThemeReleaseManifest/, 'theme package core should own release manifest normalization');
   assert.match(themeInstallServiceSource, /export function createThemeInstallService/, 'theme install service should expose an explicit service factory');
   assert.match(themeManagerDataSource, /export async function loadThemeManagerRegistry/, 'theme manager data should own registry loading');
   assert.match(themeManagerDataSource, /export async function loadThemeManagerOfficialCatalog/, 'theme manager data should own official catalog loading');
   assert.match(themeManagerDataSource, /export async function loadThemeManagerProductState/, 'theme manager data should own product-state loading');
+  assert.match(themeManagerStagingSource, /export async function stageThemeArchiveWithRuntime/, 'theme manager staging should own archive staging orchestration');
+  assert.match(themeManagerStagingSource, /export async function stageCatalogThemeWithRuntime/, 'theme manager staging should own catalog staging orchestration');
+  assert.match(themeManagerStagingSource, /export async function stageThemeUninstallWithRuntime/, 'theme manager staging should own uninstall staging orchestration');
+  assert.match(themeManagerStagingSource, /pendingSiteThemeFallback[\s\S]*setSiteThemePack/, 'theme manager staging should own site theme fallback behavior');
   assert.match(themeInstallServiceSource, /stageThemeArchive/, 'theme install service should own archive staging');
   assert.doesNotMatch(themeManagerSource, /function collectThemeArchiveEntries/, 'theme manager UI controller should not own ZIP analysis');
   assert.doesNotMatch(themeManagerSource, /function buildThemeFileChanges/, 'theme manager UI controller should not own theme file diffing');
+  assert.doesNotMatch(themeManagerSource, /function (?:setActiveSiteThemePack|stageThemeArchiveWithRuntime|stageCatalogThemeWithRuntime|stageThemeUninstallWithRuntime|stageSiteThemePack)\(/,
+    'theme manager UI controller should not re-own staging and active theme fallback rules');
   assert.doesNotMatch(themeManagerSource, /assets\/themes\/packs\.json|Official theme catalog is unavailable|Product state is unavailable|from '\.\/product-state\.js'/,
     'theme manager UI controller should not re-own registry, catalog, or product-state data loading details');
 });
@@ -222,7 +230,7 @@ await run('keeps theme manager list rendering in a view boundary', async () => {
   );
   assert.match(
     themeManagerSource,
-    /renderThemeManagerInstalledThemes\(runtime, registry, catalog, productState, \{[\s\S]*onUpdateTheme: async \(catalogEntry, entry\) => \{[\s\S]*stageCatalogThemeWithRuntime\(runtime, catalogEntry[\s\S]*onUninstallTheme: async \(entry\) => \{[\s\S]*stageThemeUninstallWithRuntime\(runtime, entry\.value\)[\s\S]*renderThemeManagerAvailableThemes\(runtime, registry, catalog, productState, \{[\s\S]*onInstallTheme: async \(entry, actionMeta = \{\}\) => \{[\s\S]*stageCatalogThemeWithRuntime\(runtime, entry,/,
+    /renderThemeManagerInstalledThemes\(runtime, registry, catalog, productState, \{[\s\S]*onUpdateTheme: async \(catalogEntry, entry\) => \{[\s\S]*stageCatalogThemeWithRuntime\(runtime, catalogEntry,[\s\S]*onUninstallTheme: async \(entry\) => \{[\s\S]*stageThemeUninstallWithRuntime\(runtime, entry\.value,[\s\S]*renderThemeManagerAvailableThemes\(runtime, registry, catalog, productState, \{[\s\S]*onInstallTheme: async \(entry, actionMeta = \{\}\) => \{[\s\S]*stageCatalogThemeWithRuntime\(runtime, entry,/,
     'theme manager controller should keep async theme staging behind view action callbacks'
   );
   assert.doesNotMatch(
